requirements.txt
discord.py==1.5.1
mysql-connector-python==8.0.22
python-dotenv==0.15.0
tabulate==0.8.7